requirements.txt
httpcore==1.0.2
httpx==0.25.1
beautifulsoup4==4.12.2
python-telegram-bot==20.6
PyYAML==6.0.1
openai>=1.3.0
